+++ CHANGELOG.md
@@ -4,11 +4,18 @@
 
 [Full Changelog](https://github.com/cloudsci/cloudmetrics/compare/v0.2.0...HEAD)
 
+*new features*
+
+- scalar field standard deviation metric, available as `scalar.std`
+  [\#64](https://github.com/cloudsci/cloudmetrics/pull/64) By Leif Denby
+  (@leifdenby)
+
 *changed defaults*
 
 - open-sky metric calculation (`mask.open_sky`) now only returns max-value by
   default [\#65](https://github.com/cloudsci/cloudmetrics/pull/65) Leif Denby
   (@leifdenby)
+
 
 ## [v0.2.0](https://github.com/cloudsci/cloudmetrics/tree/v0.2.0)